+++ src/BusinessTreeSelectBuilder/types.ts
@@ -1,28 +1,28 @@
 import type { TreeSelectProps } from 'antd';
 
-type ApiType = {
+export type ApiType = {
   /**
    * @description       获取下拉数据的接口
    * @default           '-'
    */
-  api: () => Promise<{ data: any }>;
+  api: () => Promise<any>;
   /**
    * @description       下拉选择的名字
    * @default           '-'
    */
   type: string;
   /**
-   * @description       select label的默认读取字段
+   * @description        label的默认读取字段
    * @default           'name'
    */
   labelKey?: string;
   /**
-   * @description       select value的默认读取字段
+   * @description        value的默认读取字段
    * @default           'id'
    */
   valueKey?: string;
   /**
-   * @description       设置数据不缓存
+   * @description       不缓存数据
    * @default           false
    */
   noCache?: boolean;
@@ -41,16 +41,39 @@ export type BusinessTreeSelectBuilderProps = {
   apis: ApiType[];
   /**
    * @description       默认属性;
-   * @default           '-'
+   * @default
    */
   defaultProps?: TreeSelectProps;
 };
 
-export interface BusinessTreeSelectProps<Type> extends TreeSelectProps {
+export interface BusinessTreeSelectSelfProps<Type> {
+  /**
+   * @description       类型;
+   */
   type: Type;
+  /**
+   * @description       禁用某节点;
+   */
   nodeDisabled?: (node: any) => boolean;
+  /**
+   * @description        label的默认读取字段
+   * @default           'name'
+   */
   labelKey?: string;
+  /**
+   * @description        value的默认读取字段
+   * @default           'id'
+   */
   valueKey?: string;
+  /**
+   * @description       查询接口的参数. 用法见示例
+   */
   queryParams?: Record<string, any>;
+  /**
+   * @description       不缓存数据
+   * @default           false
+   */
   noCache?: boolean;
 }
+
+export type BusinessTreeSelectProps<Type> = BusinessTreeSelectSelfProps<Type> & TreeSelectProps;